Add problem solution
The solution is in C# and can be run by creating a new Visual Studio console project and overwriting the Program.cs file that Visual Studio creates. I wasn't fully sure if a configuration file or console input was needed to the put the input and output file names by the user so I put a place for them to be entered at the top of this source file. Thanks.